backend/requirements.txt
anyio==3.6.2
asgiref==3.6.0
astroid==2.15.6
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
Django==3.2.3
django-cors-headers==4.2.0
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
drf-extra-fields==3.5.0
exceptiongroup==1.0.4
filetype==1.2.0
flake8==6.0.0
flake8-isort==6.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
iniconfig==1.1.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.0.0
platformdirs==3.8.1
pluggy==0.13.1
prettytable==3.5.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python-telegram-bot==20.1
python3-openid==3.2.0
pytz==2023.2
PyYAML==6.0
reportlab==4.0.4
requests==2.26.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.2
uritemplate==4.1.1
urllib3==1.26.15
wcwidth==0.2.5
webcolors==1.11.1
wrapt==1.15.0
django-import-export==3.2.0
gunicorn==20.1.0
